Scenario: run modal advanced settings
